Implemented AOS Interpretation
Updated RPN to handle algebraic expressions in addition to RPN
expressions.  Does validation for both RPN and AOS, conversions between
the two types, and utility function to get “canonical form” for AOS
(full_parens()).

Updated the “go out” part of the Player::manual() function to allow the
human player to enter an AOS or an RPN expression, and automatically
detect which it is.  Also provide a “bail-out” option if the human
player realizes that an equation go-out isn’t possible.

All equations are now displayed in AOS.  (Could make this a selectable
option in the future, but I suspect that AOS will always be preferred.)